requirements.txt
Flask
python-docx
PyPDF2
scikit-learn
joblib
pandas
gunicorn
